requirements.txt
streamlit
faiss-cpu
sentence-transformers
pandas==1.4.0
numpy
scikit-learn
matplotlib
plotly==5.14.1
folium
streamlit_folium
geocoder
geopy
